LLM-RAG/token-embed/requirements.txt
# Requirements for XrayGPT-like Multimodal AI System
transformers>=4.20.0
torch>=1.12.0
PyYAML>=6.0
numpy>=1.21.0

# Dependencies for image processing
Pillow>=8.0.0          # Image loading and processing
torchvision>=0.13.0    # Image transforms and preprocessing
omegaconf>=2.1.0       # Configuration management for BLIP2

# Additional dependencies for attention mechanisms and advanced features
einops>=0.6.0          # Tensor operations for attention
scipy>=1.7.0           # Scientific computing utilities
matplotlib>=3.5.0      # Visualization (optional, for attention maps)
tqdm>=4.64.0           # Progress bars
datasets>=2.0.0        # Dataset handling (optional)

# Development and testing
pytest>=7.0.0          # Testing framework (optional)
jupyter>=1.0.0         # Jupyter notebooks (optional)